Modified the Rest of DataRowService so we can get and clean most data types. Started Working on DataTable
Also added DateTime and DateTimeOffset to cleaners

Fixed in code where if I had to use a default value I wasn't letting the user know through exception that I had to do that. The Exception will not roll up unless the user asks for it. Otherwise it will be up to the user to care if it was a valid value or not by checking the Exceptions list.